Fix redundancy output

+++ include/redshow.h
@@ -64,8 +64,8 @@ typedef enum redshow_approx_level {
 typedef struct redshow_record_view {
   uint32_t function_index;
   uint64_t pc_offset;
-  uint64_t memory_id;
-  int32_t memory_op_id;
+  int32_t memory_id;
+  uint64_t memory_op_id;
   uint64_t red_count;
   uint64_t access_count;
 } redshow_record_view_t;
